Update the information related to snow at one level zone

Update the information related to snow at one level zone

stateLZ(htemp, hprecip, scaob, modelSnow, modelPrecipLZ, gca, sca, spd, wcd,
  nsno, alfa, ny)

Arguments

htemp

temperature for each elevation zone

hprecip

precipitation for each elevation zone

scaob

snow coverage observation

modelSnow

list of parameters related to snow list(nbLevelZone,unitsnow,n0,Ws,TS,CX,CFR,CGLAC,gca,UP)

modelPrecipLZ

list of parameters related to precipLZ list(nbLevelZone,Plr,hfelt,midmetp)

gca

fraction of glacier pr elevation zone, fraction of glacier covered area pr elevation zone

sca

snow coverage (NEED MORE EXPLANATION)

spd

Snow Water Equivalent in BV (NEED MORE EXPLANATION)

wcd

Free Water in BV (NEED MORE EXPLANATION)

nsno

number of event (NEED MORE EXPLANATION)

alfa

alfa parameter i nedbor gamma fordeling (NEED MORE EXPLANATION)

ny

ny parameter i nedbor gamma fordeling (NEED MORE EXPLANATION)

Examples

## Not run: ------------------------------------
# stateLZ()
## ---------------------------------------------