Update dynamical values of nu and alfa

The function varc() updates and returns dynamical values of nu and alfa, of the snowroutine. The correlation coefficient is evaluated as the sum of the correlation N*N, 2*N*u and u*u matrices Should be a decreasing value with n, large n small c All variables must be conditional (i.e. calculated on snowcovered area) Note the distinction between conditional and unconditional units of melt 18.03.2015 Recodet using only the correlation function. Alfanull ( sdalphax) and nynull( sdnux) are input

varc(ppa, ac, drange, nudyn, alphadyn, sdnux, sdalphax, nnn, u, n, redsca)

Arguments

ppa

SCA,relative reduction of SCA, previously accumulated units,current accu units and melt units

ac

(previously alfanull) parameters of relationship mean standarddeviation of conditional precip.

drange

parameters of relationship mean standarddeviation of conditional precip.

nudyn

initial distr parameters, current distr parameters

alphadyn

initial distr parameters, current distr parameters

sdnux

(ny0) initial distr parameters, current distr parameters

sdalphax

(alfa0) initial distr parameters, current distr parameters

nnn

SCA,relative reduction of SCA, previously accumulated units,current accu units and melt units

u

SCA,relative reduction of SCA, previously accumulated units,current accu units and melt units

n

SCA,relative reduction of SCA, previously accumulated units,current accu units and melt units

redsca

SCA,relative reduction of SCA, previously accumulated units,current accu units and melt units

Value

a list of three updated values: nudyn,alphadyn,nnn